// frontend/src/api/generated/models/spconsume-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { SPTransactionType } from './sptransaction-type'

/**
 * SP消費リクエストスキーマ
 * @export
 * @interface SPConsumeRequest
 */
export interface SPConsumeRequest {
  /**
   * 消費するSP量
   * @type {number}
   * @memberof SPConsumeRequest
   */
  amount: number
  /**
   * 取引の種類
   * @type {SPTransactionType}
   * @memberof SPConsumeRequest
   */
  transaction_type: SPTransactionType
  /**
   * 取引の説明
   * @type {string}
   * @memberof SPConsumeRequest
   */
  description: string
  /**
   *
   * @type {string}
   * @memberof SPConsumeRequest
   */
  related_entity_type?: string | null
  /**
   *
   * @type {string}
   * @memberof SPConsumeRequest
   */
  related_entity_id?: string | null
  /**
   * 追加のメタデータ
   * @type {object}
   * @memberof SPConsumeRequest
   */
  metadata?: object
}